Lagt til egen service for sammenligning og lagt til for KLP
