+++ src/main/kotlin/no/nav/tjenestepensjon/simulering/v2025/tjenestepensjon/v1/service/TpUtil.kt
@@ -1,19 +1,11 @@
 package no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service
 
-import io.github.oshai.kotlinlogging.KotlinLogging
 import no.nav.tjenestepensjon.simulering.common.AlderUtil.bestemUttaksalderVedDato
-import no.nav.tjenestepensjon.simulering.model.domain.pen.FremtidigInntekt
-import no.nav.tjenestepensjon.simulering.model.domain.pen.SimulerAFPOffentligLivsvarigRequest
-import no.nav.tjenestepensjon.simulering.v2025.afp.v1.AFPOffentligLivsvarigSimuleringService
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.domain.Maanedsutbetaling
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.domain.Utbetalingsperiode
-import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.dto.request.SimulerTjenestepensjonRequestDto
-import org.springframework.stereotype.Component
 import java.time.LocalDate
 
-@Component
-class TpUtil(private val afp: AFPOffentligLivsvarigSimuleringService) {
-    private val log = KotlinLogging.logger {}
+object TpUtil {
     fun grupperMedDatoFra(utbetalingsliste: List<Utbetalingsperiode>, foedselsdato: LocalDate): List<Maanedsutbetaling> {
         return utbetalingsliste
             .groupBy { it.fom }
@@ -23,30 +15,4 @@ class TpUtil(private val afp: AFPOffentligLivsvarigSimuleringService) {
             }
             .sortedBy { it.fraOgMedDato }
     }
-
-    fun sammenlignOgLoggAfp(request: SimulerTjenestepensjonRequestDto, utbetalingsperiode: List<Utbetalingsperiode>) {
-        val afpLokal = afp.simuler(
-            SimulerAFPOffentligLivsvarigRequest(
-            fnr = request.pid,
-            fom = request.uttaksdato,
-            fodselsdato = request.foedselsdato,
-            fremtidigeInntekter = listOf(
-                opprettNaaverendeInntektFoerUttak(request),
-                FremtidigInntekt(
-                    0,
-                    request.uttaksdato
-                )
-            ),
-            )
-        )
-        val afpFraTpOrdning = utbetalingsperiode.filter { it.ytelseType == "OAFP" }
-        log.info { "AFP fra Tp ordning: $afpFraTpOrdning \n AFP fra lokal $afpLokal" }
-    }
-
-    private fun opprettNaaverendeInntektFoerUttak(request: SimulerTjenestepensjonRequestDto) = FremtidigInntekt(
-        request.sisteInntekt,
-        fjorAarSomManglerOpptjeningIPopp()
-    )
-
-    private fun fjorAarSomManglerOpptjeningIPopp(): LocalDate = LocalDate.now().minusYears(1).withDayOfYear(1)
 }
