+++ src/main/kotlin/no/nav/tjenestepensjon/simulering/v2025/tjenestepensjon/v1/service/spk/SPKTjenestepensjonClient.kt
@@ -8,9 +8,9 @@ import no.nav.tjenestepensjon.simulering.sporingslogg.SporingsloggService
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.domain.SimulertTjenestepensjon
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.dto.request.SimulerTjenestepensjonRequestDto
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.exception.TjenestepensjonSimuleringException
+import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service.SammenlignAFPService
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service.TjenestepensjonV2025Client
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service.TjenestepensjonV2025Client.Companion.TJENESTE
-import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service.TpUtil
 import no.nav.tjenestepensjon.simulering.v2025.tjenestepensjon.v1.service.spk.dto.SPKSimulerTjenestepensjonResponse
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
@@ -22,7 +22,7 @@ import org.springframework.web.reactive.function.client.bodyToMono
 class SPKTjenestepensjonClient(
     private val spkWebClient: WebClient,
     private val sporingsloggService: SporingsloggService,
-    private val TpUtil: TpUtil
+    private val sammenlign: SammenlignAFPService
 ) : TjenestepensjonV2025Client, Pingable {
     private val log = KotlinLogging.logger {}
 
@@ -38,7 +38,9 @@ class SPKTjenestepensjonClient(
                 .retrieve()
                 .bodyToMono<SPKSimulerTjenestepensjonResponse>()
                 .block()
-            return response?.let { Result.success(SPKMapper.mapToResponse(it, dto).also { res -> TpUtil.sammenlignOgLoggAfp(request, res.utbetalingsperioder) }) } ?: Result.failure(TjenestepensjonSimuleringException("No response body"))
+            return response?.let { Result.success(SPKMapper.mapToResponse(it, dto)
+                .also { res -> sammenlign.sammenlignOgLoggAfp(request, res.utbetalingsperioder) }) }
+                ?: Result.failure(TjenestepensjonSimuleringException("No response body"))
         } catch (e: WebClientResponseException) {
             val errorMsg = "Failed to simulate tjenestepensjon 2025 hos SPK ${e.responseBodyAsString}"
             log.error(e) { errorMsg }
